+++ code/statements/loops-examples/reset.c
@@ -3,12 +3,12 @@
 /*@
   requires \valid(array + (0 .. length-1));
   assigns  array[0 .. length-1];
-  ensures  \forall size_t j; 0 <= j < length ==> array[j] == 0;
+  ensures  \forall integer j; 0 <= j < length ==> array[j] == 0;
 */
 void reset(int* array, size_t length){
   /*@
     loop invariant 0 <= i <= length;
-    loop invariant \forall size_t j; 0 <= j < i ==> array[j] == 0;
+    loop invariant \forall integer j; 0 <= j < i ==> array[j] == 0;
     loop assigns i, array[0 .. length-1];
     loop variant length-i;
   */
